+++ lib/reco.rb
@@ -1,5 +1,6 @@
 require 'optparse'
 require 'forwardable'
+require 'ostruct'
 # require 'fileutils'
 require 'time'
 require 'tempfile'
@@ -9,9 +10,15 @@ require 'tempfile'
 def load_modules(*preload_modules)
   STDERR.printf "==> loading modules: " if @debug_autoload
   loadpath = File.join(File.dirname(__FILE__), File.basename(__FILE__, '.rb'), '') # include empty slash!
-  basemodules = preload_modules.collect {|f| loadpath+f.to_s+'.rb'}
-  allmodules = Dir[loadpath + '**.rb']
-  (basemodules+(allmodules-basemodules)).each do |file|
+  basemodules = preload_modules.collect do |f|
+    file = loadpath+f.to_s+'.rb'
+    STDERR.printf("%s, ", f.to_s) if @debug_autoload
+    load file
+    file
+  end
+  include Reco
+  allmodules = Dir[loadpath + '**.rb'] - basemodules
+  allmodules.each do |file|
     STDERR.printf "%s, ", File.basename(file,'.rb') if @debug_autoload
     load file
   end
@@ -19,6 +26,5 @@ def load_modules(*preload_modules)
 end
 
 load_modules :version, :helper
-include Reco # @todo do this in helper