+++ History.md
@@ -1,4 +1,4 @@
 ### 0.0.1
 
 * Skeletonizing...
-
+* ported some brittle helper stuff from emigma project.